Define dependencies using pipreqs
