+++ requirements.txt
@@ -1,0 +1,4 @@
+pyautogui==0.9.53
+pyperclip==1.8.2
+PySimpleGUI==4.59.0
+requests==2.27.1
