+++ test_util.py
@@ -1,0 +1,5 @@
+from util import containsAll
+
+
+def test_containsAll():
+    assert containsAll('Hello world!','world')
